requirements.txt
Package  Version
-------- -------
asgiref  3.10.0
Django   5.2.7
pillow   12.0.0
pip      24.3.1
sqlparse 0.5.3
tzdata   2025.2